#68 Make FindPwContents  that  is  separated from FindPw
