+++ src/pages/FindPw/style.module.scss
@@ -22,7 +22,7 @@
     width: 32px;
   }
 }
-.inner {
+.contents {
   position: relative;
   width: 100%;
   //100vh - bottomNav.height = topBar.height
